Implemented recurrent neural network, now only with RNN layers
